week-02-vit-image-classifier/requirements.txt
torch
transformers==4.40.0
pillow
pandas
